requirements.txt
# Core ML/Audio Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
librosa>=0.10.0
scipy>=1.10.0
soundfile>=0.12.0

# Deep Learning (optional for advanced models)
torch>=2.0.0
torchaudio>=2.0.0

# Data Handling
datasets>=2.12.0
huggingface_hub>=0.15.0

# API
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.65.0
joblib>=1.3.0

# Development
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytest>=7.4.0

# AI Voice Generation APIs (for data collection)
openai>=1.0.0
elevenlabs>=0.2.0
